Split OcrUtilities, move the OCR text assembly into Core.Windows
Batch 4c of the Core split, plus the 3d RTL blocker it depended on.

The split went the direction the call-site census pointed rather than the
one the file's size suggests. Tests/OcrTests.cs alone held 43 of
OcrUtilities' ~80 references and every one is against the text-assembly
subset, so that half keeps the OcrUtilities name and moves to
Text-Grab.Core.Windows: GetTextFromOcrLine, FilterFurigana,
FilterFuriganaLines, OrderLinesForReadingFlow, BuildTextFromOcrLines,
ShouldUseParagraphDetection, GroupWrappedParagraphLines, UnionRectangles,
IsWrappedLine, IsWrappedParagraph, GetStringFromOcrOutputs,
ParseOcrResultIntoWordBorderInfos, the nested PositionedOcrLine and
GroupedOcrLines, and the SpaceJoiningWordRegex it all runs on. OcrTests
therefore needed no edit for those, and rides into Tests.Core.Windows for
free in 7a.

The app-coupled half took the new name OcrSourceUtilities - screen and
window capture, engine dispatch, file and BitmapSource sources. Its ~35
references across nine files were renamed. It still needs WPF throughout,
and its engine dispatch needs WindowsAiUtilities and LanguageUtilities,
neither of which has moved.

The moved half's four settings reads (CorrectErrors, CorrectToLatin,
ParagraphDetection, RemoveFurigana) were already on ITextGrabSettings, so
this needed no interface addition - the static DefaultSettings field
became SettingsAccess.Current reads, same as 4a and 4b.

Deferred, as section 4.4 said to: LoadBitmapFromFile builds a WPF
BitmapImage to apply EXIF rotation, so it and its two callers
(OcrAbsoluteFilePathAsync, OcrFile) stay in the app. Section 7 updated.

BuildTextFromOcrLines calls language.IsRightToLeft(), which 3d had left
app-side because XmlLanguage comes from PresentationCore. Settled the
behaviour question with a throwaway WPF probe rather than by reasoning:
XmlLanguage.GetLanguage(tag).GetEquivalentCulture().TextInfo.IsRightToLeft
and CultureInfo.GetCultureInfo(tag).TextInfo.IsRightToLeft agreed on all
24 tags probed - ar, ar-EG, ar-SA, he, he-IL, ur, ur-PK, fa, fa-IR, ckb,
ps-AF, sd-Arab-PK, yi, he-Hebr-IL, ar-XX, en, en-US, ja, zh-Hans, de-DE,
and the unresolvable xx, xx-YY, und and "". So the ILanguage overload
moved into Core.Windows LanguageExtensions with a CultureInfo lookup,
guarded by a CultureNotFoundException catch returning false (XmlLanguage
fell back to the invariant culture, which is LTR). That left the Language
overload with zero call sites - all five live IsRightToLeft calls are on
ILanguage - so Extensions/LanguageRtlExtensions.cs is deleted outright.

Deleted GetBoundingRect(this OcrLine) as section 8 dead code, re-verified
first: WinRtOcrLinesWords has its own private GetBoundingRect and
UIAutomationUtilities calls the unrelated TextPatternRange
GetBoundingRectangles.

Self-assembly-reflection audit: no GetExecutingAssembly/GetEntryAssembly/
GetCallingAssembly, typeof(x).Assembly, Assembly.GetName/.Location/
.CodeBase, AssemblyMetadata, AppContext.BaseDirectory or manifest-resource
lookups in any of the three files.

Also recorded in the plan: 4e cannot precede 5a. LanguageService touches
InputLanguageManager in exactly one private method and is otherwise WinRT,
so the whole class can move behind a registered resolver - but
GetAllLanguages and GetOCRLanguage call WindowsAiUtilities, which is
deferred on SoftwareBitmapExtensions in 5a.

Gate: dotnet build Tests/Tests.csproj -c Debug -p:Platform=x64 - clean.
Tests.Core 514, Tests.Core.Windows 3, Tests 993 passed/7 skipped -
identical to baseline, total 1510 conserved.
